+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 owncloud-news (4.1.4)
+* **Enhancement**: Add admin setting to set a custom explore service URL
 * **Enhancement**: Add explore button and show explore button on startup
 * **Enhancement**: Show a hint when no articles are available